[SELC-6226] feat: check for email mismatch and refactor form validation
