+++ src/locale/it.ts
@@ -246,6 +246,8 @@ export default {
           'A causa di un errore del sistema non è possibile completare <1 />la procedura. Ti chiediamo di riprovare più tardi.',
         backHome: 'Torna alla home',
         moreInformationOnRoles: 'Vorrei più informazioni',
+        requiredError: 'Campo obbligatorio',
+        duplicateTaxCodeDifferentEmail: 'L’indirizzo email inserito è diverso da quello comunicato in precedenza'
       },
       backButton: 'Indietro',
       continueButton: 'Continua',
