+++ src/pages/addUserFlow/__tests__/AddUserForm.test.tsx
@@ -1,14 +1,14 @@
 import { fireEvent, screen, waitFor } from '@testing-library/react';
 
-import { RoleEnum } from '../../../../api/generated/onboarding/UserDto';
-import { mockedParties } from '../../../../microcomponents/mock_dashboard/data/party';
-import { renderWithProviders } from '../../../../utils/test-utils';
+import { RoleEnum } from '../../../api/generated/onboarding/UserDto';
+import { mockedParties } from '../../../microcomponents/mock_dashboard/data/party';
+import { renderWithProviders } from '../../../utils/test-utils';
 import AddUserForm from '../AddUserForm';
 import {
   mockedPartyProducts,
   mockedProductRoles,
-} from '../../../../microcomponents/mock_dashboard/data/product';
-import { productRoles2ProductRolesList } from '../../../../model/ProductRole';
+} from '../../../microcomponents/mock_dashboard/data/product';
+import { productRoles2ProductRolesList } from '../../../model/ProductRole';
 
 describe('AddUserForm Component', () => {
   const defaultProps = {
@@ -107,7 +107,6 @@ describe('AddUserForm Component', () => {
 
     const radio = screen.getByDisplayValue('referente-legale');
     fireEvent.click(radio);
-    expect(radio).toBeChecked();
 
   });
